requirements.txt
Flask>=2.0.1
Flask-SocketIO>=5.1.1
eventlet>=0.33.0
pillow>=10.0.0
requests>=2.31.0 